TW-109: separate chat_list_view to seperate components
